reorganize repository a bit
